chore: Restructure packages for vssNodes, vssPaths and connection
