+++ app/src/main/java/org/eclipse/kuksa/testapp/databroker/JavaDataBrokerEngine.java
@@ -38,8 +38,8 @@ import org.eclipse.kuksa.connectivity.databroker.response.VssNodeUpdateResponse;
 import org.eclipse.kuksa.coroutine.CoroutineCallback;
 import org.eclipse.kuksa.proto.v1.KuksaValV1.GetResponse;
 import org.eclipse.kuksa.proto.v1.KuksaValV1.SetResponse;
-import org.eclipse.kuksa.testapp.databroker.factory.DataBrokerConnectorFactory;
-import org.eclipse.kuksa.testapp.databroker.model.ConnectionInfo;
+import org.eclipse.kuksa.testapp.databroker.connection.factory.DataBrokerConnectorFactory;
+import org.eclipse.kuksa.testapp.databroker.connection.model.ConnectionInfo;
 import org.eclipse.kuksa.vsscore.model.VssNode;
 
 import java.util.HashSet;
