+++ app/src/main/kotlin/org/eclipse/kuksa/testapp/databroker/connection/view/connection/TlsOptionsView.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,36 +14,42 @@
  * limitations under the License.
  *
  * SPDX-License-Identifier: Apache-2.0
- *
  */
 
-package org.eclipse.kuksa.testapp.databroker.view.connection
+package org.eclipse.kuksa.testapp.databroker.connection.view.connection
 
 import android.net.Uri
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.text.KeyboardActionScope
+import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.constraintlayout.compose.ConstraintLayout
-import org.eclipse.kuksa.testapp.databroker.model.ConnectionInfo
+import org.eclipse.kuksa.testapp.databroker.connection.model.ConnectionInfo
 import org.eclipse.kuksa.testapp.databroker.view.DefaultEdgePadding
+import org.eclipse.kuksa.testapp.databroker.view.DefaultElementPadding
 import org.eclipse.kuksa.testapp.databroker.view.FileSelectorSettingView
 import org.eclipse.kuksa.testapp.extension.fetchFileName
 
 @Composable
-fun AuthenticationOptionsView(
+fun TlsOptionsView(
     connectionInfo: ConnectionInfo,
     modifier: Modifier = Modifier,
-    onAuthenticationStateChanged: (Boolean) -> Unit = {},
-    onJwtSelected: (Uri) -> Unit = {},
+    onTlsStateChanged: (Boolean) -> Unit = {},
+    onCertificateSelected: (Uri) -> Unit = {},
+    onAuthorityOverrideChanged: (String) -> Unit = {},
+    onKeyboardDone: (KeyboardActionScope) -> Unit = {},
 ) {
     val context = LocalContext.current
 
@@ -52,20 +58,20 @@ fun AuthenticationOptionsView(
             modifier = Modifier
                 .fillMaxWidth()
                 .clickable {
-                    onAuthenticationStateChanged(!connectionInfo.isAuthenticationEnabled)
+                    onTlsStateChanged(!connectionInfo.isTlsEnabled)
                 },
         ) {
             val (text, checkbox) = createRefs()
             Text(
-                text = "Authentication",
+                text = "TLS",
                 modifier = Modifier.constrainAs(text) {
                     top.linkTo(parent.top)
                     bottom.linkTo(parent.bottom)
                 },
             )
             Checkbox(
-                checked = connectionInfo.isAuthenticationEnabled,
-                onCheckedChange = onAuthenticationStateChanged,
+                checked = connectionInfo.isTlsEnabled,
+                onCheckedChange = onTlsStateChanged,
                 modifier = Modifier
                     .constrainAs(checkbox) {
                         top.linkTo(parent.top)
@@ -76,15 +82,31 @@ fun AuthenticationOptionsView(
             )
         }
 
-        if (connectionInfo.isAuthenticationEnabled) {
-            val uri = Uri.parse(connectionInfo.jwtUriPath ?: "")
-            val fileName = uri.fetchFileName(context) ?: "Select JWT..."
+        if (connectionInfo.isTlsEnabled) {
+            val uri = connectionInfo.certificate.uri
+            val fileName = uri.fetchFileName(context) ?: "Select certificate..."
 
             FileSelectorSettingView(
-                label = "JWT",
+                label = "Certificate",
                 value = fileName,
-                onResult = onJwtSelected,
+                onResult = onCertificateSelected,
                 modifier = Modifier.padding(end = 10.dp),
+            )
+
+            Spacer(modifier = Modifier.padding(top = DefaultElementPadding))
+
+            TextField(
+                value = connectionInfo.certificate.overrideAuthority,
+                onValueChange = onAuthorityOverrideChanged,
+                keyboardActions = KeyboardActions(
+                    onDone = onKeyboardDone,
+                ),
+                modifier = Modifier
+                    .fillMaxWidth(),
+                singleLine = true,
+                label = {
+                    Text(text = "Authority override")
+                },
             )
         }
     }
@@ -93,19 +115,19 @@ fun AuthenticationOptionsView(
 @Preview
 @Composable
 private fun DisabledTlsOptionsViewPreview() {
-    val connectionInfo = ConnectionInfo(isAuthenticationEnabled = false)
+    val connectionInfo = ConnectionInfo(isTlsEnabled = false)
 
     Surface {
-        AuthenticationOptionsView(connectionInfo)
+        TlsOptionsView(connectionInfo)
     }
 }
 
 @Preview
 @Composable
 private fun EnabledTlsOptionsViewPreview() {
-    val connectionInfo = ConnectionInfo(isAuthenticationEnabled = true)
+    val connectionInfo = ConnectionInfo(isTlsEnabled = true)
 
     Surface {
-        AuthenticationOptionsView(connectionInfo)
+        TlsOptionsView(connectionInfo)
     }
 }
