+++ app/src/test/kotlin/org/eclipse/kuksa/testapp/model/ConnectionInfoTest.kt
@@ -24,9 +24,9 @@ import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.eclipse.kuksa.test.kotest.Unit
-import org.eclipse.kuksa.testapp.databroker.model.Certificate
-import org.eclipse.kuksa.testapp.databroker.model.ConnectionInfo
-import org.eclipse.kuksa.testapp.databroker.model.ConnectionInfoSerializer
+import org.eclipse.kuksa.testapp.databroker.connection.model.Certificate
+import org.eclipse.kuksa.testapp.databroker.connection.model.ConnectionInfo
+import org.eclipse.kuksa.testapp.databroker.connection.model.ConnectionInfoSerializer
 import java.io.File
 import java.io.FileWriter
